backend/requirements.txt
# OBM RecFind Drop-In Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
asyncpg>=0.31.0
alembic==1.13.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiohttp==3.9.1
playwright>=1.57.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pydantic>=2.12.5
pydantic-settings==2.1.0
cachetools==5.3.2
apscheduler==3.10.4
structlog==24.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
